refactor(axum): split the proxy handlers from their wire shapes
The bind-conflict fix pushed `handlers/proxy.rs` from 387 to 399 lines, and
the file-size ratchet is there precisely so a file already over budget cannot
keep growing.

The seam is obvious once looked for. `ProxyStatus`, `StartProxyConfig` and
`StartPinnedBody` with their two conversions are one thing — the rules for
resolving what a caller left out, which is where the surprises live and which
every test in the file was already exercising — and the routing is another.
`wire` takes the shapes and all nine tests; `mod` keeps the handlers and the
idempotency decisions, and drops to 148 lines.

`to_api_status` also loses a duplicated arm: `Stopped` and `Crashed` built the
same all-`None` response separately.
