+++ crates/gglib-axum/src/handlers/proxy/wire.rs
@@ -1,17 +1,18 @@
-//! Proxy handlers - OpenAI-compatible proxy management.
+//! The proxy endpoints' request and response shapes, and their conversions.
+//!
+//! Split from the handlers so the resolution rules below — which are where the
+//! surprises live, since an omitted field means "use what is configured"
+//! rather than "use the compile-time default" — can be read and tested without
+//! the routing around them.
 
-use axum::{Json, extract::State};
-
-use crate::{error::HttpError, state::AppState};
 use gglib_app_services::types::AppSettings;
-use gglib_core::ports::AppEventEmitter;
 use gglib_core::server_config::{ServerConfigOptions, resolve_context_size};
 use gglib_core::settings::DEFAULT_PROXY_PORT;
 use gglib_runtime::proxy::ProxyConfig as RuntimeProxyConfig;
 use gglib_runtime::proxy::ProxyStatus as RuntimeProxyStatus;
 
 /// Proxy status response.
-/// Matches Tauri's ProxyStatus for frontend compatibility.
+/// Matches Tauri's `ProxyStatus` for frontend compatibility.
 #[derive(Debug, Clone, serde::Serialize)]
 pub struct ProxyStatus {
     pub running: bool,
@@ -62,16 +63,25 @@ pub struct StartProxyConfig {
     pub allowed_hosts: Vec<String>,
 }
 
-/// Convert runtime ProxyStatus to API ProxyStatus.
-fn to_api_status(s: RuntimeProxyStatus, pinned_model: Option<String>) -> ProxyStatus {
+/// Request body for `POST /api/proxy/start-pinned`.
+///
+/// The GUI names a model and its overrides; the daemon runs the same
+/// cascade as `gglib serve` (`gglib_app_services::launch_options`) so the
+/// two surfaces cannot drift. `options` uses the camelCase wire form of the
+/// bare `/api/servers/start` body; `proxy` the snake_case form of
+/// `/api/proxy/start`.
+#[derive(Debug, Clone, serde::Deserialize)]
+pub struct StartPinnedBody {
+    pub model_id: i64,
+    #[serde(default)]
+    pub options: gglib_app_services::types::StartServerRequest,
+    #[serde(default)]
+    pub proxy: StartProxyConfig,
+}
+
+/// Convert runtime `ProxyStatus` to API `ProxyStatus`.
+pub fn to_api_status(s: RuntimeProxyStatus, pinned_model: Option<String>) -> ProxyStatus {
     match s {
-        RuntimeProxyStatus::Stopped => ProxyStatus {
-            running: false,
-            port: None,
-            current_model: None,
-            model_port: None,
-            pinned_model: None,
-        },
         RuntimeProxyStatus::Running { address } => ProxyStatus {
             running: true,
             port: Some(address.port()),
@@ -79,7 +89,9 @@ fn to_api_status(s: RuntimeProxyStatus, pinned_model: Option<String>) -> ProxySt
             model_port: None,
             pinned_model,
         },
-        RuntimeProxyStatus::Crashed => ProxyStatus {
+        // A crashed proxy is not serving, so it reports exactly as a stopped
+        // one does: no port to hand out, and no pin still in force.
+        RuntimeProxyStatus::Stopped | RuntimeProxyStatus::Crashed => ProxyStatus {
             running: false,
             port: None,
             current_model: None,
@@ -87,13 +99,6 @@ fn to_api_status(s: RuntimeProxyStatus, pinned_model: Option<String>) -> ProxySt
             pinned_model: None,
         },
     }
-}
-
-/// Fetch current proxy status from backend.
-async fn fetch_status(state: &AppState) -> ProxyStatus {
-    let s = state.proxy.status().await;
-    let pinned = state.proxy.pinned_model();
-    to_api_status(s, pinned)
 }
 
 /// Convert handler config to runtime config, resolving anything omitted from
@@ -105,7 +110,7 @@ async fn fetch_status(state: &AppState) -> ProxyStatus {
 /// `ProxyOps::ensure_running`. Going straight to `DEFAULT_PROXY_PORT` here
 /// would silently ignore a changed `proxy_port` for every client of this
 /// endpoint.
-fn to_runtime_config(cfg: &StartProxyConfig, settings: &AppSettings) -> RuntimeProxyConfig {
+pub fn to_runtime_config(cfg: &StartProxyConfig, settings: &AppSettings) -> RuntimeProxyConfig {
     let default_context = resolve_context_size(&ServerConfigOptions {
         context_size: cfg.default_context,
         global_default_ctx: settings.default_context_size,
@@ -139,148 +144,6 @@ fn to_runtime_config(cfg: &StartProxyConfig, settings: &AppSettings) -> RuntimeP
         api_key: cfg.api_key.clone(),
         allowed_hosts: cfg.allowed_hosts.clone(),
     }
-}
-
-/// Request body for `POST /api/proxy/start-pinned`.
-///
-/// The GUI names a model and its overrides; the daemon runs the same
-/// cascade as `gglib serve` (`gglib_app_services::launch_options`) so the
-/// two surfaces cannot drift. `options` uses the camelCase wire form of the
-/// bare `/api/servers/start` body; `proxy` the snake_case form of
-/// `/api/proxy/start`.
-#[derive(Debug, Clone, serde::Deserialize)]
-pub struct StartPinnedBody {
-    pub model_id: i64,
-    #[serde(default)]
-    pub options: gglib_app_services::types::StartServerRequest,
-    #[serde(default)]
-    pub proxy: StartProxyConfig,
-}
-
-/// Start the proxy pinned to one model, resolving the launch cascade
-/// server-side — the GUI counterpart of `gglib serve`.
-pub async fn start_pinned(
-    State(state): State<AppState>,
-    Json(body): Json<StartPinnedBody>,
-) -> Result<Json<ProxyStatus>, HttpError> {
-    let globals = gglib_app_services::launch_options::ProxyGlobals {
-        host: body.proxy.host.clone(),
-        proxy_port: body.proxy.port,
-        llama_base_port: body.proxy.llama_base_port,
-        default_ctx: body.proxy.default_context,
-        cache_enabled: body.proxy.cache.unwrap_or(false),
-        slot_dir: body.proxy.slot_dir.clone(),
-        api_key: body.proxy.api_key.clone(),
-        allowed_hosts: body.proxy.allowed_hosts.clone(),
-    };
-
-    let plan = state
-        .proxy
-        .plan_pinned(body.model_id, &body.options, globals)
-        .await?;
-
-    // Delegate to the ordinary start path with the planned pin. Slot dir and
-    // default context ride the cascade exactly as the CLI sends them: the
-    // master switch has been applied and the context fully resolved.
-    let proxy_config = plan.unified.to_proxy_config();
-    let cfg = StartProxyConfig {
-        pinned: Some(plan.pinned),
-        slot_dir: proxy_config.slot_dir,
-        default_context: Some(proxy_config.default_context),
-        // Sampling rides the pinned model's launch options, exactly as the
-        // CLI sends it — never a proxy-wide override.
-        inference_override: None,
-        ..body.proxy
-    };
-    start(State(state), Json(Some(cfg))).await
-}
-
-/// Get current proxy status.
-pub async fn status(State(state): State<AppState>) -> Json<ProxyStatus> {
-    Json(fetch_status(&state).await)
-}
-
-/// Start the proxy (idempotent).
-pub async fn start(
-    State(state): State<AppState>,
-    Json(cfg): Json<Option<StartProxyConfig>>,
-) -> Result<Json<ProxyStatus>, HttpError> {
-    let cfg = cfg.unwrap_or_default();
-
-    // Resolve context size through the shared 3-level fallback chain
-    // (flag > settings default > hard-coded default), matching CLI behavior.
-    let settings = state.settings.get().await?;
-    let runtime_cfg = to_runtime_config(&cfg, &settings);
-
-    // Idempotent: if already running (Conflict), treat as success — unless
-    // the caller asked for a pinned mode the running proxy does not match,
-    // where "success" would silently hand them an endpoint without the
-    // refuse-foreign-models guarantee they explicitly requested.
-    match state.proxy.start(runtime_cfg, cfg.pinned.clone()).await {
-        Ok(_addr) => {}
-        Err(e) => {
-            let http: HttpError = e.into();
-            if !matches!(http, HttpError::Conflict(_)) {
-                return Err(http);
-            }
-
-            // `Conflict` covers two different outcomes. `AlreadyRunning` is
-            // the idempotent success this branch exists for; `BindFailed` is
-            // somebody else's process on the port, and it leaves nothing
-            // running. Only the first is success, so ask which one happened
-            // rather than assuming — treating a bind failure as success
-            // answered 200 with `running: false` and discarded the one message
-            // that named the port and what to do about it.
-            if !fetch_status(&state).await.running {
-                return Err(http);
-            }
-
-            let requested = cfg.pinned.as_ref().map(|p| p.name.clone());
-            if requested.is_some() && state.proxy.pinned_model() != requested {
-                return Err(HttpError::Conflict(format!(
-                    "the proxy is already running {} — stop it first (`gglib proxy stop`)",
-                    match state.proxy.pinned_model() {
-                        Some(name) => format!("pinned to '{name}'"),
-                        None => "unpinned".to_string(),
-                    }
-                )));
-            }
-        }
-    }
-
-    let status = fetch_status(&state).await;
-
-    // Emit proxy started event if proxy is now running
-    if status.running
-        && let Some(port) = status.port
-    {
-        state
-            .sse
-            .emit(gglib_core::events::AppEvent::proxy_started(port));
-    }
-
-    Ok(Json(status))
-}
-
-/// Stop the proxy (idempotent).
-pub async fn stop(State(state): State<AppState>) -> Result<Json<ProxyStatus>, HttpError> {
-    // Idempotent: if not running (Conflict), treat as success
-    match state.proxy.stop().await {
-        Ok(()) => {
-            // Emit proxy stopped event on clean shutdown
-            state
-                .sse
-                .emit(gglib_core::events::AppEvent::proxy_stopped());
-        }
-        Err(e) => {
-            let http: HttpError = e.into();
-            if !matches!(http, HttpError::Conflict(_)) {
-                return Err(http);
-            }
-        }
-    }
-
-    Ok(Json(fetch_status(&state).await))
 }
 
 #[cfg(test)]
@@ -395,5 +258,16 @@ mod tests {
         };
         let runtime_cfg = to_runtime_config(&cfg, &settings);
         assert_eq!(runtime_cfg.default_context, 16_384);
+    }
+
+    /// A crashed proxy must not look reachable. It reports stopped, with no
+    /// port for a client to be handed and no pin left standing.
+    #[test]
+    fn a_crashed_proxy_reports_as_stopped() {
+        let status = to_api_status(RuntimeProxyStatus::Crashed, Some("qwen".to_owned()));
+
+        assert!(!status.running);
+        assert_eq!(status.port, None);
+        assert_eq!(status.pinned_model, None);
     }
 }
